# src/GroundControl.Api/OpenApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GroundControl.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/groups/{id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get a group",
        "description": "Returns a group by its unique identifier. Includes an ETag header for optimistic concurrency.",
        "operationId": "GetGroupHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Update a group",
        "description": "Updates an existing group. Requires an If-Match header with the current ETag value.",
        "operationId": "UpdateGroupHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Delete a group",
        "description": "Deletes a group if it has no dependents. Requires an If-Match header.",
        "operationId": "DeleteGroupHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{id}/members": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "List group members",
        "description": "Returns all users that are members of the specified group.",
        "operationId": "ListGroupMembersHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{id}/members/{userId}": {
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Add or update a group member",
        "description": "Adds a user to the group or updates their membership role.",
        "operationId": "SetGroupMemberHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetGroupMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Remove a group member",
        "description": "Removes a user from the specified group.",
        "operationId": "RemoveGroupMemberHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/personal-access-tokens/{id}": {
      "get": {
        "tags": [
          "PersonalAccessTokens"
        ],
        "summary": "Get a personal access token",
        "description": "Returns metadata for a personal access token owned by the current user.",
        "operationId": "GetPatHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PersonalAccessTokens"
        ],
        "summary": "Revoke a personal access token",
        "description": "Permanently revokes a personal access token owned by the current user.",
        "operationId": "RevokePatHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GrantDto": {
        "required": [
          "roleId"
        ],
        "type": "object",
        "properties": {
          "resource": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional group resource targeted by the grant.",
            "format": "uuid"
          },
          "roleId": {
            "type": "string",
            "description": "Gets the identifier of the granted role.",
            "format": "uuid"
          },
          "conditions": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Gets optional scope value conditions that further restrict access."
          }
        },
        "description": "Represents a role grant in request and response contracts."
      },
      "GroupResponse": {
        "required": [
          "id",
          "name",
          "version",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the unique identifier for the group.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets the display name for the group."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional human-readable description for the group."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Gets the optimistic concurrency version.",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "Gets the timestamp when the group was created.",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "description": "Gets the identifier of the user that created the group.",
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets the timestamp when the group was last updated.",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "description": "Gets the identifier of the user that last updated the group.",
            "format": "uuid"
          }
        },
        "description": "Represents the API response body for a group."
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PatResponse": {
        "required": [
          "id",
          "name",
          "tokenPrefix",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tokenPrefix": {
            "type": "string"
          },
          "permissions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "isRevoked": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SetGroupMemberRequest": {
        "required": [
          "roleId"
        ],
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "description": "Gets the identifier of the role to grant for this group membership.",
            "format": "uuid"
          }
        },
        "description": "Represents the request body for adding a user as a group member with a specific role."
      },
...
      "UpdateGroupRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "Gets the display name for the group."
          },
          "description": {
            "maxLength": 500,
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional human-readable description for the group."
          }
        },
        "description": "Represents the request body for updating a group."
      },
...
      "UserResponse": {
        "required": [
          "id",
          "username",
          "email",
          "grants",
          "isActive",
          "version",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the unique identifier for the user.",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "description": "Gets the username."
          },
          "email": {
            "type": "string",
            "description": "Gets the email address."
          },
          "grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantDto"
            },
            "description": "Gets the role grants assigned to the user."
          },
          "isActive": {
            "type": "boolean",
            "description": "Gets a value indicating whether the user is active."
          },
          "externalProvider": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the external identity provider name."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Gets the optimistic concurrency version.",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "Gets the timestamp when the user was created.",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "description": "Gets the identifier of the user that created this user.",
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets the timestamp when the user was last updated.",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "description": "Gets the identifier of the user that last updated this user.",
            "format": "uuid"
          }
        },
        "description": "Represents the API response body for a user."
      },
...
    }
  },
...
}